requirements.txt
sentence-transformers
pinecone-client
python-dotenv
google-genai